Update GetGrainInfo.m
Added one line to make sure grainID is a vector, not a rectangular matrix.  Will obviously break for hex-grid and L-grid scans, but I'm not even sure if HKL does those?  Also, we might consider adding in the ability to run find grains for .ang files if we don't have/trust the OIM grain file.