Update SplitDD.m
This missing negative sign arose from the mistaken belief that the X-Y coordinates from .ang and .ctf files are in the reference frame of the plot, but they're actually in their own "spatial" reference frame.  This negative sign counteracts the fact that Fa is calculated based on a negative change in the y direction.  It might be simpler just to change it there...