Changed the default PC in case the cpr file doesnt have one
When you export tiffs from an AZtec data file (as opposed to producing images as you scan), the cpr file contains no PC and you need to get it from the tiffs in PC calibration.  This can be annoying when checking your ROI settings, so I changed the default PC to something more reasonable.

This is a band-aid, logically we would check the tiff here, but for that we need to make sure we already had an image loaded.  You can't optimize pattern center from an arbitrary point very well, meaning that when you get this option you're kind of stuck with reading from a tiff or manually changing these values.  If you have time, Brian, the best course would be to check the tiff and if all else fails, have them enter the value manually.