Fixed PC cal for real patterns
If you try to do PC calibration for real patterns it fails thanks to the new dynamic code.  Fixed it with an otherwise statement.  Now whenever you're planning a dynamic pattern run, your calibration will alos be dynamic, otherwise you use kinematic simulated patterns.

Ideally, we should allow the user to pick a different form of simulated pattern for calibration, so this can be viewed as a temporary fix.