Fixed an error reading Fit for zero solutions in HKL
AZtecHKL treats zero solutions differently than OIM in regards to the Fit parameter (called MAD by AZtec).  For a zero solution, AZtec sets the value to zero, making it look very good when it is actually terrible.  This makes selecting a reference pattern via the usual methods problematic, as it will always choose these terrible points, or most of the time.  I added a line that changes all zero MADs to 180.  Seems to work