Added a slight pause in EditRefPoints
For reasons not entirely clear (possibly due to different version of matlab, or simply quirks of different computers), ginputs tend to malfunction when the focus moves to a different axis, which was causing the middle button click to show a pattern feature when editing reference points to glitch in some cases.  Adding a slight pause fixes this problem for those that had it, and is roughly undetectable for those who didn't.