Added in compatibility with HKL's camera Euler angles.
Currently, the code doesn't allow you to arbitrarily set the camera and sample orientations.  This makes getting the precise geometry to line up problematic.  This update allows the user to read the camera Euler angles from tiff files when looking at an AZtec scan.  To see this benefit most clearly, use the test geometry feature.

The way to activate this is painfully ad hoc, whenever HREBSDPrep is run it just pops up a dialog box whether you'd like to read the camera Euler angles in for from the tiffs or just rely on the user supplied orientation.  This is intended as a stand in until the problem gets a proper GUI solution.

This change should be completely invisible when using an OIM scan.

-T